day 19 part 1
works with my cooked input but i couldn't get a working solution that worked for both blueprint 1 and blueprint 2 in the example

ended up trying a bunch of solutions and changing back and forth between both of them to make it work. so sad lol

took this solution that worked for 2 and attempted to run it and it works :(

idk why. have already spent too much time on this.